backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
# python-multipart
pydantic==2.8.2
sentence-transformers==4.1.0
faiss-cpu==1.11.0
PyMuPDF==1.26.3
openpyxl==3.1.5
requests==2.32.3
python-dotenv==0.21.0